+++ src/main/java/src/exceptions/EmptyResultException.java
@@ -4,6 +4,6 @@ import org.springframework.dao.EmptyResultDataAccessException;
 
 public class EmptyResultException extends EmptyResultDataAccessException {
 
-    public EmptyResultException(Long id, String type) { super(String.format("%s with id %s not found", type, id), id.intValue()); }
+    public EmptyResultException(Integer id, String type) { super(String.format("%s with id %s not found", type, id), id.intValue()); }
 
 }
