+++ src/main/java/src/model/ApartmentsRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ApartmentsRepository extends JpaRepository<Apartment, Long> {
+public interface ApartmentsRepository extends JpaRepository<Apartment, Integer> {
 
     List<Apartment> findByType(String type);
