+++ src/main/java/src/model/InvestmentRepository.java
@@ -3,5 +3,5 @@ package src.model;
 import org.springframework.data.jpa.repository.JpaRepository;
 import src.factory.Investment;
 
-public interface InvestmentRepository extends JpaRepository<Investment, Long> {
+public interface InvestmentRepository extends JpaRepository<Investment, Integer> {
 }
